backend/requirements.txt
agno
fastapi
uvicorn
python-dotenv
openai
supabase
